requirements.txt
# Core dependencies
python-telegram-bot>=20.0,<22.0
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0

# Production dependencies
aiohttp>=3.9.0  # For healthcheck HTTP server
psutil>=5.9.0  # For system metrics

# Optional but recommended
ta>=0.11.0  # Technical Analysis library
